+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## V.6.9
+## LAST
 
 * The namespace of the RichText class has been changed to avadim\FastExcelWriter\RichText
+* Deprecated methods removed: Sheet::setColStyles(), Sheet::setColOptions(), Sheet::getExternalLinks(), Sheet::setPageOptions(), Sheet::setRowOptions(), Sheet::setRowStyles(),
+* An instance of the Style object can be used as an argument in all methods where an array with styles was passed
 
 ## V.6.8